Un par de cosas movidas fuera de Utils.java a sus respectivos métodos
